fix: not being able to shift click items in the crafting grid
Fixes #922

+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/grid/SnapshotExtractTransaction.java
@@ -46,7 +46,7 @@ public class SnapshotExtractTransaction implements ExtractTransaction {
         if (!available.contains(craftingMatrixResource)) {
             final long amount = rootStorage.extract(
                 craftingMatrixResource,
-                Long.MAX_VALUE,
+                Integer.MAX_VALUE, // integer for NeoForge support
                 Action.SIMULATE,
                 playerActor
             );
